Create AreaCircleAccurate
You are given a positive integer A denoting the radius of a circle. You have to calculate the area of the circle.

NOTE: Since, the answer can be a real number, you have to return the ceil value of the area. Ceil value of a real number X is the smallest integer that is greater than or equal to X.



Problem Constraints

1 <= A <= 1000



Input Format

First and only argument is an integer A.



Output Format

Return a single integer denoting the ceil value of area of the circle.



Example Input

Input 1:

 A = 1 
Input 2:

 A = 4 


Example Output

Output 1:

 4 
Output 2:

 51