Create Sunny number
we need only to check whether N+1 is the perfect square or not.

Read or initialize a number (num).
Add 1 to the given number i.e. num+1.
Find the square root of num+1.
If the square root is an integer, the given number is sunny, else not a sunny number.